+++ project/src/main/java/dao/RestaurantDao.java
@@ -99,7 +99,7 @@ public class RestaurantDao {
 		template.getMapper(cls).deleteMenu(num);
 	}
 
-	public int pointAvg(int num) {
+	public double pointAvg(int num) {
 		return template.getMapper(cls).pointAvg(num);
 //	}
